🐛 不必要なコードの削除と仮データの移行 #4
