+++ src/app/page.tsx
@@ -1,15 +1,12 @@
-import { AuthorModel, PostCardModel, RestaurantModel } from "@/model/PostCard";
+import {
+  AuthorModel,
+  Mockdata,
+  PostCardModel,
+  RestaurantModel,
+} from "@/model/PostCard";
 import PostCard from "src/app/components/Post/Card";
 import Navbar from "./components/elements/Navbar";
 import Link from "next/link";
-
-// PostCardに渡すデータ(仮)
-export const Mockdata = {
-  author: AuthorModel,
-  post: PostCardModel,
-  restaurant: RestaurantModel,
-  isLike: true,
-};
 
 export default function Home() {
   return (
